+++ docs/14_0_0/gettingstarted/whatsnew.md
@@ -27,6 +27,10 @@ Look into [migration guide](https://primefaces.github.io/primefaces/14_0_0/#/../
     * Added `icon` and `iconPos` to allow icons to be used in badge.
     * Added 'onclick` to allow client side click of badge. If not set it delegates the click to its target element.
 
+* CSV - Client Side Validation / CommandButton
+    * Added `enabledByValidateClient` to enable CommandButton after CSV was successful (used together with `<p:clientValidator event="..." />` on all relevant input-components)  
+      See https://primefaces.github.io/primefaces/14_0_0/#/core/csv?id=immediate-csv
+
 * DataExporter
     * Added `bufferSize` to control how many items are fetched at a time when `DataTable#lazy` is enabled.